backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
httpx==0.26.0

# File uploads
python-multipart==0.0.6
aiofiles==23.2.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Production server
gunicorn==21.2.0
